move paramkind
